refactored to use databsase loader
